Update ChevronButton position/colour
